requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.0
redis==5.0.3
